Add a UMBaseConanFile for reusable code
the _um_data function is  the most efficient function, since we can't
easily cache and lazy-compute the data, since that requires and
`__init__` method to be written for the recipe which uses this. I
think that would be overkill, considering this is only called when
installing / creating.

Contributes to CURA-9365