Add compiler/linker args and remove `.c` extensions
Update, I believe I found a bug in Sip I will post a question /report on the PyQt mailinglist later today.
Our Python binding projects (pynest2d, pyArcus, pySavitar) are all written in the c++17 standard. When Sip generates
the binding files it will create .cpp files from our .sip specification files. These .cpp files along with the .c and
.h files that consist of the Sip API are then copied to the build folder. In order to compile the Python module with
c++17 we also need to pass the argument -std=c++17 to the compiler. The pyproject.toml file allows for us to parse
these extra compiler arguments. But these arguments are used for creating each object file .o also the object files
generated from pure C .c files. Most compilers (GNU GCC, MSVC, LLVM Clang) are smart enough to discard that argument,
print out a warning and continue with compiling. But not Apple Clang. This compiler treats it as an error.

I believe it is not necessary to add the .c Sip API files to the list of sources, Since they are already in the chain
when the headers are included. Commenting out the following lines fixes the issue and allows us to compile with
Apple Clang again. https://www.riverbankcomputing.com/hg/sip/file/tip/sipbuild/module/module.py#l103

This will still copy the .c files but not add them to the list of source.

Since modifying those lines is actually patching Sip and that would also require us to download, patch and install a
modified Sip version. I will create a temporary workaround. I will add a custom builder or bindings class which removes
.c files from the list of sources before they're passed to the Extension module of Setuptools. Not the nicest solution.
But it show the clearest intent and reason why we do this. We can then wait on an answer from Riverbanking if this is
indeed a bug and if it will be fixed.

Contributes to CURA-8788